+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Removed `DEFINED_STATUS_CODES` constant.
 - Removed `response.json()` calls internally. Request functions return `Response` object now.
 - Added `searchParams()` method to add additional URLSearchParams.
+- Fixed `addResponse()` interceptor issues.
 
 ## [2.1.0 (2024-09-28)](https://github.com/axe-api/axe-api/compare/2.1.0...2.0.1)